sys/v4l2/gstv4l2: skip set caps processing if the parameters are unchanged
v4l2src re-creates the buffer pool in set_caps() when the buffers
already have been allocated. If always-copy is set to false,
the buffer dequeued from V4L2 driver can be passed to next plugin
without copying. Therefore, set_caps() could free the buffer which
is in use.

set_caps() may be invoked by core libraries even if there is no change
to the caps value. This behavior could free the buffer which is in use
by another plugin. This patch skips the set_caps() processing if the caps
parameters are unchanged.