sys/v4l2/gstv4l2: return an error from set_caps() if trying to re-create in-use buffers
Whether the caps can be changed depends on the always-copy option.

If the caps of a buffer are changed, the buffer must be
re-created because the buffer size may change. If always-copy is
set to false, a buffer which is in use could be freed so that it
can be re-created.

For this reason, if always-copy is set to false, changing the caps is
not supported and this patch returns an error.