avidemux: set frame start code to VC-1 advanced profile stream
VC-1 advanced profile constrains the bitstream format to pair
the frame data with the frame start code.
This patch pulls frame data from previous plugin with extra field
for containing a start code and inserts the start code before the
frame data if the stream doesn't has the start code.